feat: implement global exception handling

+++ src/main/java/com/shitanshu/shopping/service/ProductService.java
@@ -3,6 +3,7 @@ import com.shitanshu.shopping.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import java.util.ArrayList;
 import java.util.List;
+import com.shitanshu.shopping.exception.ProductNotFoundException;
 
 import org.springframework.stereotype.Service;
 
@@ -23,9 +24,12 @@ public class ProductService {
         
         return productRepository.findAll();
     }
-    // New Method
     public Product getProductById(int id) {
-        return productRepository.findById(id).orElse(null);
+
+        return productRepository.findById(id)
+                .orElseThrow(() ->
+                        new ProductNotFoundException("Product with ID " + id + " not found"));
+
     }
     
     public Product addProduct(Product product) {
@@ -34,28 +38,28 @@ public class ProductService {
     
     public Product updateProduct(int id, Product updatedProduct) {
 
-        Product existingProduct = productRepository.findById(id).orElse(null);
+        Product existingProduct = productRepository.findById(id)
+                .orElseThrow(() ->
+                        new ProductNotFoundException("Product with ID " + id + " not found"));
 
-        if (existingProduct != null) {
+        existingProduct.setName(updatedProduct.getName());
+        existingProduct.setBrand(updatedProduct.getBrand());
+        existingProduct.setDescription(updatedProduct.getDescription());
+        existingProduct.setPrice(updatedProduct.getPrice());
+        existingProduct.setOldPrice(updatedProduct.getOldPrice());
+        existingProduct.setRating(updatedProduct.getRating());
+        existingProduct.setImage(updatedProduct.getImage());
+        existingProduct.setCategory(updatedProduct.getCategory());
+        existingProduct.setStock(updatedProduct.getStock());
 
-            existingProduct.setName(updatedProduct.getName());
-            existingProduct.setBrand(updatedProduct.getBrand());
-            existingProduct.setDescription(updatedProduct.getDescription());
-            existingProduct.setPrice(updatedProduct.getPrice());
-            existingProduct.setOldPrice(updatedProduct.getOldPrice());
-            existingProduct.setRating(updatedProduct.getRating());
-            existingProduct.setImage(updatedProduct.getImage());
-            existingProduct.setCategory(updatedProduct.getCategory());
-            existingProduct.setStock(updatedProduct.getStock());
-
-            return productRepository.save(existingProduct);
-        }
-
-        return null;
+        return productRepository.save(existingProduct);
     }
     public void deleteProduct(int id) {
 
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ProductNotFoundException("Product with ID " + id + " not found");
+        }
 
+        productRepository.deleteById(id);
     }
 }
